requirements/base.txt
django
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter